+++ src/sbt-test/sbt-yourkit/simple/Main.scala
@@ -2,6 +2,6 @@ package co.vitaler.sbt.yourkit.test
 
 object Main {
   def main(args: Array[String]): Unit = {
-    println("main() of application called")
+    println(s"main() of application called with args: ${args.mkString(",")}")
   }
 }