+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased]
 
+## [0.1.2] - 2019-07-24
+
 ### Changed
 
 - Changed the type of the startup options setting to be a `Map[String, String]`
@@ -27,6 +29,7 @@
 
 - Initial version imported from upstream
 
-[Unreleased]: https://github.com/vital-software/sbt-update-lines/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/vital-software/sbt-update-lines/compare/v0.1.2...HEAD
+[0.1.2]: https://github.com/vital-software/sbt-update-lines/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/vital-software/sbt-update-lines/compare/v0.0.3...v0.1.1
 [0.0.3]: https://github.com/vital-software/sbt-update-lines/releases/tag/v0.0.3